+++ .travis.yml
@@ -11,9 +11,9 @@ install:
 before_script:
 - "export DISPLAY=:99.0"
 - "sh -e /etc/init.d/xvfb start"
-- sleep 3 # give xvfb some time to start
-- yarn start & # start a Web server
-- sleep 3 # give Web server some time to bind to sockets, etc 
+- sleep 3
+- yarn start &
+- sleep 3 
 script:
 - yarn lint
 - yarn test:coverage
