+++ README.md
@@ -13,6 +13,13 @@ yarn start
 # Please read *package.json* for more available options.
 ```
 
+## Tests
+>Please note, that in order to execute perceptual diff tests, you need to boot an application on your local machine (port:3000 by default). All the screenshots will be saved at tmp/niffy directory.
+```bash
+# to run tests simply type yarn test:coverage in your terminal
+yarn test:coverage
+```
+
 ## Description
 Collections is a web application, which displays images of various topics in a convenient way. Images is preseneted as a gallery (grid) of tiles of a different width and height. User has an option to click on a specific image to see it in a full sized mode. User also has an option to search and filter images.
 
@@ -25,6 +32,7 @@ API used: [unsplash.com](https://unsplash.com/documentation).
 * Mobx-React
 * Glamorous
 * Jest
+* Nightmarejs and Niffy
 * Webpack
 
 ## Demo
