+++ src/__tests__/e2e_tests/Gallery.e2e.js
@@ -1,4 +1,3 @@
-import Nightmare from 'nightmare';
 import Niffy from 'niffy';
 import co from 'co';
 import {
@@ -6,35 +5,14 @@ import {
   NIFFY_TEST_HOST,
   NIFFY_OPTIONS,
   ASYNC_CALLBACK_TIMEOUT,
+  MOBILE_SCREEN_WIDTH,
+  MOBILE_SCREEN_HEIGHT,
 } from '../../../test-setup';
 
-describe('Image Gallery', () => {
-  it(
-    'show open a modal window, if user clicks on an image',
-    async () => {
-      const nightmare = Nightmare(NIFFY_OPTIONS);
-      await nightmare
-        .goto(NIFFY_TEST_HOST)
-        .wait(5000)
-        .click('img')
-        .wait('#preview-modal')
-        .exists('#preview-modal')
-        .visible('#preview-modal')
-        .evaluate(() => document.querySelector('#preview-modal img'))
-        .end()
-        .then((image) => {
-          expect(image).toHaveProperty('src');
-          expect(image).toHaveProperty('complete', true);
-        })
-        .catch((error) => {
-          console.error(`An error has occurred: ${error}`);
-        });
-    },
-    ASYNC_CALLBACK_TIMEOUT,
-  );
+describe('Visual Regression Tests', () => {
+  let niffy;
 
-  describe('Image Gallery perceptual diff test', () => {
-    let niffy;
+  describe('Gallery', () => {
     beforeAll(() => {
       niffy = new Niffy(NIFFY_BASE_HOST, NIFFY_TEST_HOST, NIFFY_OPTIONS);
     });
@@ -45,22 +23,36 @@ describe('Image Gallery', () => {
       });
     });
 
-    it(
-      'should open a modal window with an image',
-      async () => {
-        await co(function* () {
-          return yield niffy.test('/', async (nightmare) => {
-            await co(function* () {
-              return yield nightmare
-                .click('img')
-                .wait('#preview-modal')
-                .wait(3000)
-                .end();
+    describe('Gallery perceptual diff test', () => {
+      it(
+        'Desktop: should display a grid with images',
+        async () => {
+          await co(function* () {
+            return yield niffy.test('/', async (nightmare) => {
+              await co(function* () {
+                return yield nightmare.wait(3000);
+              });
             });
           });
-        });
-      },
-      ASYNC_CALLBACK_TIMEOUT,
-    );
+        },
+        ASYNC_CALLBACK_TIMEOUT,
+      );
+
+      it(
+        'Mobile: should display a grid with images',
+        async () => {
+          await co(function* () {
+            return yield niffy.test('/', async (nightmare) => {
+              await co(function* () {
+                return yield nightmare
+                  .viewport(MOBILE_SCREEN_WIDTH, MOBILE_SCREEN_HEIGHT)
+                  .wait(3000);
+              });
+            });
+          });
+        },
+        ASYNC_CALLBACK_TIMEOUT,
+      );
+    });
   });
 });
